requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning (Optional - for future enhancements)
tensorflow>=2.13.0  # Optional for neural network enhancements
keras>=2.13.0       # Optional for autoencoder

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0      # For interactive plots

# Data Processing
imbalanced-learn>=0.11.0  # For SMOTE and advanced balancing techniques

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# CLI and Utilities
click>=8.1.0        # For command-line interfaces
tqdm>=4.65.0        # Progress bars
colorama>=0.4.6     # Colored terminal output

# Logging
loguru>=0.7.0       # Advanced logging

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.3.0       # Code formatter
flake8>=6.0.0       # Linter
pylint>=2.17.0      # Code analysis
mypy>=1.3.0         # Type checking

# Documentation
sphinx>=7.0.0       # Documentation generator
sphinx-rtd-theme>=1.2.0

# Model Serving (Optional - for production deployment)
flask>=2.3.0        # Web framework
flask-cors>=4.0.0   # CORS support
gunicorn>=20.1.0    # WSGI server
fastapi>=0.100.0    # Alternative API framework
uvicorn>=0.22.0     # ASGI server

# Model Monitoring
mlflow>=2.4.0       # ML experiment tracking (optional)

# Network Analysis (Optional - for real network integration)
scapy>=2.5.0        # Packet manipulation
pyshark>=0.6.0      # Packet analysis
